+++ src/audio_core/mix_context.cpp
@@ -2,6 +2,8 @@
 // Licensed under GPLv2 or any later version
 // Refer to the license.txt file included.
 
+#include <algorithm>
+
 #include "audio_core/behavior_info.h"
 #include "audio_core/common.h"
 #include "audio_core/effect_context.h"
