+++ src/common/logging/text_formatter.cpp
@@ -13,6 +13,7 @@
 #include "common/common_funcs.h"
 #include "common/logging/filter.h"
 #include "common/logging/log.h"
+#include "common/logging/log_entry.h"
 #include "common/logging/text_formatter.h"
 #include "common/string_util.h"
