+++ src/common/logging/types.h
@@ -4,8 +4,6 @@
 
 #pragma once
 
-#include <chrono>
-
 #include "common/common_types.h"
 
 namespace Common::Log {
@@ -129,21 +127,6 @@ enum class Class : u8 {
     Network,           ///< Network emulation
     WebService,        ///< Interface to yuzu Web Services
     Count              ///< Total number of logging classes
-};
-
-/**
- * A log entry. Log entries are store in a structured format to permit more varied output
- * formatting on different frontends, as well as facilitating filtering and aggregation.
- */
-struct Entry {
-    std::chrono::microseconds timestamp;
-    Class log_class{};
-    Level log_level{};
-    const char* filename = nullptr;
-    unsigned int line_num = 0;
-    std::string function;
-    std::string message;
-    bool final_entry = false;
 };
 
 } // namespace Common::Log
