+++ src/core/hle/service/lbl/lbl.cpp
@@ -2,6 +2,7 @@
 // Licensed under GPLv2 or any later version
 // Refer to the license.txt file included.
 
+#include <cmath>
 #include <memory>
 
 #include "common/logging/log.h"
